requirements.txt
tensorflow
keras
numpy
matplotlib
pillow
scikit-learn
opencv-python
jupyter
gradio
streamlit
seaborn